[chore] add user enum and move dir position
