+++ src/domain/user/repository/agreements.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '@common/database/prisma.service';
-import { AgreementCategory } from '@domain/user.enum';
+import { AgreementCategory } from '@domain/user/user.enum';
 
 @Injectable()
 export class AgreementRepository {
